+++ src/app/column-stacked/column-stacked.demo.ts
@@ -7,6 +7,7 @@ import {
 import {
   SohoRadarComponent
 } from 'ids-enterprise-ng';
+import {SohoColumnComponent} from '../../soho/column';
 
 @Component({
   selector: 'soho-column-demo',
@@ -14,6 +15,9 @@ import {
 })
 export class ColumnStackedDemoComponent implements OnInit {
 
+  @ViewChild(SohoColumnComponent) sohoColumnComponent: SohoColumnComponent;
+
+  private selection: SohoColumnSelected  = {fieldName: 'name', fieldValue: 'Jul'};
   public columnStackedData = [{
     data: [{
       name: 'Jan', value: 12,
@@ -101,4 +105,14 @@ export class ColumnStackedDemoComponent implements OnInit {
 
   constructor() {}
   ngOnInit() {}
+
+  setChartSelection() {
+    const SohoColumnSelected: SohoColumnSelected = this.selection;
+    this.sohoColumnComponent.setSelected(SohoColumnSelected);
+  }
+
+  toggleChartSelection () {
+    const SohoColumnSelected: SohoColumnSelected = this.selection;
+    this.sohoColumnComponent.toggleSelected(SohoColumnSelected);
+  }
 }