+++ src/soho/radar/soho-radar.d.ts
@@ -106,16 +106,21 @@ interface SohoRadar {
   updated(settings?: SohoRadarOptions): void;
 
   /** Gets the currently selected element */
-  getSelected(o: SohoRadarOptions, isToggle?: boolean);
+  getSelected(o?: SohoRadarOptions, isToggle?: boolean);
 
   /** Sets the currently selected element */
-  setSelected();
+  setSelected(selected: SohoRadarSelected);
 
   /** Toggles the currently selected element */
-  toggleSelected();
+  toggleSelected(selected: SohoRadarSelected);
 
   /** Destroys the control on completion. */
   destroy(): void;
+}
+
+interface SohoRadarSelected {
+  fieldName: string;
+  fieldValue: any;
 }
 
 interface JQuery {